+++ src/main/java/ru/vyarus/guice/persist/orient/db/PoolManager.java
@@ -6,13 +6,10 @@ import com.google.inject.Provider;
  * Manage pool of connections for specific connection type (e.g. object connections).
  * Pool also works as connection provider, but implementation should not allow to obtain connection outside transaction
  * (unit of work).
- * <p>While transaction manager defines global unit of work scope, each pool manager defines
+ * <p>While transaction manager defines global unit of work scope, each pool manager may define
  * its own thread local transaction.
- * It is impossible in orient to share transaction between different connections (by design),
- * but most likely in most cases only one pool will be used
- * in transaction.</p>
- * <p>To change pool size use {@code OGlobalConfiguration.DB_POOL_MIN.setValue()} and
- * {@code OGlobalConfiguration.DB_POOL_MAX.setValue()}. By default pools it's 1-20.</p>
+ * In default pool implementations single document connection is shared between pools.</p>
+ * <p>To change pool size use {@code OGlobalConfiguration.DB_POOL_MAX.setValue()}. By default pools it's 100.</p>
  * <p>Do not use global pools in implementations - always create new one to avoid possible
  * collisions with other pools</p>
  * <p>Implementation must rely on {@link ru.vyarus.guice.persist.orient.db.user.UserManager} for actual
@@ -25,7 +22,9 @@ import com.google.inject.Provider;
 public interface PoolManager<T> extends Provider<T> {
 
     /**
-     * Start pool. Will be called by PersistService implementation..
+     * Start pool. Will be called by PersistService implementation.
+     * Method is responsible for initialization, specific for connection type (this mean new no tx connection
+     * must be created to let orient configure database for specific connection type (required for object and graph)
      *
      * @param uri  database uri
      */