+++ src/main/java/ru/vyarus/guice/persist/orient/finder/scanner/InvisibleForScanner.java
@@ -11,8 +11,12 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
  *
  * @author Vyacheslav Rusakov
  * @since 17.10.2014
+ * @deprecated finders now completely
+ * <a href="https://github.com/xvik/guice-ext-annotations#usage">controlled by guice</a>
+ * you can use @ProvidedBy instead of classpath scanning
  */
 @Target(TYPE)
 @Retention(RUNTIME)
+@Deprecated
 public @interface InvisibleForScanner {
 }